requirements.txt
aiofiles==25.1.0
aiogram==3.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp_socks==0.11.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
attrs==25.4.0
build==1.4.2
CacheControl==0.14.4
certifi==2026.2.25
charset-normalizer==3.4.5
cleo==2.1.0
colorama==0.4.6
crashtest==0.4.1
distlib==0.4.0
dulwich==1.1.0
fastjsonschema==2.21.2
filelock==3.25.2
findpython==0.7.1
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
keyring==25.7.0
magic-filter==1.0.12
more-itertools==11.0.1
msgpack==1.1.2
multidict==6.7.1
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pbs-installer==2026.3.25
pkginfo==1.12.1.2
platformdirs==4.9.4
poetry-core==2.3.2
propcache==0.4.1
psycopg==3.3.3
pydantic==2.12.5
pydantic_core==2.41.5
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-discovery==1.2.1
python-dotenv==1.2.2
python-socks==2.8.1
pywin32-ctypes==0.2.3
RapidFuzz==3.14.3
redis==7.4.0
requests==2.32.5
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.49
tomlkit==0.14.0
trove-classifiers==2026.1.14.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
virtualenv==21.2.0
yarl==1.23.0
zstandard==0.25.0
